+++ packages/data-schemas/src/models/balance.ts
@@ -1,0 +1,6 @@
+import mongoose from 'mongoose';
+import balanceSchema from '~/schema/balance';
+import type * as t from '~/types';
+
+export const Balance =
+  mongoose.models.Balance || mongoose.model<t.IBalance>('Balance', balanceSchema);
