+++ packages/data-schemas/src/types/convo.ts
@@ -1,0 +1,53 @@
+import type { Document, Types } from 'mongoose';
+
+// @ts-ignore
+export interface IConversation extends Document {
+  conversationId: string;
+  title?: string;
+  user?: string;
+  messages?: Types.ObjectId[];
+  agentOptions?: unknown;
+  // Fields provided by conversationPreset (adjust types as needed)
+  endpoint?: string;
+  endpointType?: string;
+  model?: string;
+  region?: string;
+  chatGptLabel?: string;
+  examples?: unknown[];
+  modelLabel?: string;
+  promptPrefix?: string;
+  temperature?: number;
+  top_p?: number;
+  topP?: number;
+  topK?: number;
+  maxOutputTokens?: number;
+  maxTokens?: number;
+  presence_penalty?: number;
+  frequency_penalty?: number;
+  file_ids?: string[];
+  resendImages?: boolean;
+  promptCache?: boolean;
+  thinking?: boolean;
+  thinkingBudget?: number;
+  system?: string;
+  resendFiles?: boolean;
+  imageDetail?: string;
+  agent_id?: string;
+  assistant_id?: string;
+  instructions?: string;
+  stop?: string[];
+  isArchived?: boolean;
+  iconURL?: string;
+  greeting?: string;
+  spec?: string;
+  tags?: string[];
+  tools?: string[];
+  maxContextTokens?: number;
+  max_tokens?: number;
+  reasoning_effort?: string;
+  // Additional fields
+  files?: string[];
+  expiredAt?: Date;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
