+++ packages/data-schemas/src/types/index.ts
@@ -1,8 +1,6 @@
-// User types
 export * from './user';
-
-// Session types
-export * from './session';
-
-// Token types
 export * from './token';
+export * from './convo';
+export * from './session';
+export * from './balance';
+export * from './message';
